casino_ai_api/requirements.txt
fastapi==0.115.0
uvicorn==0.32.1
httpx==0.27.2
pydantic==2.10.3
python-dotenv==1.0.1
psycopg2-binary==2.9.10
